project restructure for dockerization
