+++ backend/api/Dockerfile
@@ -1,0 +1,9 @@
+FROM python:3.7-alpine
+
+WORKDIR /code
+
+COPY requirements.txt requirements.txt
+RUN pip install -r requirements.txt
+
+COPY . .
+CMD ["python3.7", "manage.py", "runserver"]
