+++ backend/edicoBeidnakerfi/urls.py
@@ -20,5 +20,5 @@ from django.urls import path
 urlpatterns = [
     path('admin/', admin.site.urls),
 
-    url(r'', include("backend.urls"))
+    url(r'', include("api.urls"))
 ]
